+++ NEWS.md
@@ -1,5 +1,5 @@
 ## 3.0.0-SNAPSHOT 2025-mm-dd
-* [MODAUD-XXX](https://folio-org.atlassian.net/browse/MODAUD-XXX) - Description
+* [MODAUD-264](https://folio-org.atlassian.net/browse/MODAUD-264) - Respect tenant timezone when showing loan due date in log record
 
 ## 2.11.1 2025-04-15 
 * [MODAUD-250](https://folio-org.atlassian.net/browse/MODAUD-250) - Version history of "MARC" records is not tracked